Added htpasswd.py
Basic authentication requires a htpasswd file containing user names
and passwords. This file is compatible with the corresponding file
used by the Apache web server.
Typically is created using the program 'htpasswd' that comes with
the Apache web server.
As we do not want to be dependent on additional software packages,
htpasswd.py was added to provide an independent way to create
and manage htpasswd files.